chore: use benchmarks as package folder
