+++ package.json
@@ -50,7 +50,6 @@
     "test:coverage": "jest --coverage --coverageReporters=lcov --coverageReporters=text-summary",
     "type-check": "tsc --noEmit",
     "type-check:watch": "tsc --noEmit --watch",
-    "benchmark": "ts-node benchmarks/run-all.ts",
     "commit": "cz",
     "format": "prettier --write \"**/*.ts\"",
     "lint": "eslint lib/**/*.ts  __tests__/**/*.ts scripts/**/*.ts",
@@ -75,7 +74,6 @@
     "@types/jest": "29.5.14",
     "@typescript-eslint/eslint-plugin": "8.29.1",
     "@typescript-eslint/parser": "8.29.1",
-    "benny": "^3.7.1",
     "cz-conventional-changelog": "^3.3.0",
     "eslint": "9.24.0",
     "eslint-config-prettier": "10.1.1",
@@ -91,7 +89,6 @@
     "rimraf": "6.0.1",
     "semantic-release": "^25.0.2",
     "ts-jest": "29.3.1",
-    "ts-node": "^10.9.2",
     "tsyringe": "^4.10.0",
     "typescript": "5.8.3"
   },
